move scanner to its own module
